refactor(types): matching new project structure
