--- src/types/general.ts
@@ -1,3 +1,0 @@
-export type LintStagedConfig = {
-  [key: string]: string[];
-};
