+++ src/types/project.ts
@@ -1,0 +1,3 @@
+import { Project } from 'projen';
+
+export type IsProjectType<T> = T extends Project ? T : never;
